requirements.txt
aiosqlite==0.21.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cryptography==46.0.1
cssselect2==0.7.0
dnspython==2.8.0
dominate==2.9.1
email-validator==2.3.0
fastapi==0.115.12
flasgger==0.9.7.1
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.2
fonttools==4.55.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.2
pillow==10.4.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
pyphen==0.15.0
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.5
rpds-py==0.22.3
sendgrid==6.12.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.46.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.0
visitor==0.1.3
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
zopfli==0.2.3.post1
